services/worker/requirements.txt
nats-py
redis
prometheus_client